+++ CHANGELOG.md
@@ -3,3 +3,8 @@ All notable changes to `laravel-ns-api` will be documented in this file
 
 ## 1.0.0 - 2020-05-05
 Initial setup
+
+## 1.0.1 - 2020-10-04
+ - Add `Product` and `RouteStation` entities.
+ - Update the name of the fields being used on the NS API for entity `Departure`.
+ - Add new method `getDelay()` to calculate the actual delay of the departure.